chore: simplify worker serve for e2e tests and move test cases
