+++ packages/playground/server-components/tests/worker-e2e.test.ts
@@ -1,5 +1,5 @@
 import {port} from './worker-e2e.serve';
-import testCases from '../../e2e-test-cases';
+import testCases from './e2e-test-cases';
 
 describe('Worker', () => {
   // @ts-ignore
